Moved setup to new module
